Registering all the binary functions.
