--- pkg/sql/plan/function2/builtinSet_binary_arjun.go
@@ -1,78 +1,0 @@
-// Copyright 2021 - 2022 Matrix Origin
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package function2
-
-import (
-	"github.com/matrixorigin/matrixone/pkg/container/vector"
-	"github.com/matrixorigin/matrixone/pkg/vectorize/timediff"
-	"github.com/matrixorigin/matrixone/pkg/vm/process"
-)
-
-// ENDSWITH
-
-func EndsWith(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-// EXTRACT
-
-func ExtractFromDatetime(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-func ExtractFromDate(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-func ExtractFromTime(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-func ExtractFromVarchar(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-// FINDINSET
-
-func FindInSet(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-// INSTR
-
-func Instr(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-// LEFT
-
-func Left(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-// STARTSWITH
-
-func Startswith(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-// TIMEDIFF
-
-func TimeDiff[T timediff.DiffT](ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
-
-//POW
-
-func Power(ivecs []*vector.Vector, result vector.FunctionResultWrapper, _ *process.Process, length int) error {
-	return nil
-}
